+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [1.3.0](https://github.com/MikeTaylor/mod-app-manager/tree/v1.3.0) (IN PROGRESS)
 
 * Update the JSON Schemas and examples for FAM files to match the current [file-format definition](https://github.com/MikeTaylor/mafia/blob/master/doc/folio-app-metadata.md). Fixes #9.
+* Support `_tenant` interface for `loadReference` and `loadSample` operations. Fixes #3.
+* Use `id` from mod-configuration as the app-source ID. Fixes #11.
 
 ## [1.2.0](https://github.com/MikeTaylor/mod-app-manager/tree/v1.2.0) (2022-08-10)